Modify in UI design
Many Changes in design and add new columns in two tables in database

+++ database/migrations/2025_03_26_142453_create_about_statics_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
 
             $table->string('image');
-            $table->tinyInteger('number');
+            $table->mediumInteger('number');
             $table->string('description');
             $table->char('after_number', 3)->nullable();
             $table->char('before_number', 3)->nullable();
